scrape/requirements.txt
scrapy
google-cloud-storage
google-crc32c
kfp
google-cloud-aiplatform
